🔨 [services] Move function `loadtemplate` from `generate` to `loadtemplate`
