+++ src/cutty/services/loadtemplate.py
@@ -1,0 +1,22 @@
+"""Loading templates."""
+import pathlib
+from typing import Optional
+
+import platformdirs
+
+from cutty.filesystems.domain.purepath import PurePath
+from cutty.repositories.adapters.storage import getdefaultrepositoryprovider
+from cutty.repositories.domain.repository import Repository as Template
+
+
+def loadtemplate(
+    template: str, checkout: Optional[str], directory: Optional[pathlib.PurePosixPath]
+) -> Template:
+    """Load a template repository."""
+    cachedir = pathlib.Path(platformdirs.user_cache_dir("cutty"))
+    repositoryprovider = getdefaultrepositoryprovider(cachedir)
+    return repositoryprovider(
+        template,
+        revision=checkout,
+        directory=(PurePath(*directory.parts) if directory is not None else None),
+    )
